+++ source/core/wifi_multidoc_webconfig.c
@@ -573,7 +573,7 @@ static int update_vap_info_managed_guest(void *data, void *amenities_blob, wifi_
         vap_info->u.bss_info.enabled = true;
         vap_info->u.bss_info.bssMaxSta = 75;
         memset(&vap_info->u.bss_info.security.repurposed_radius,0,sizeof(vap_info->u.bss_info.security.repurposed_radius));
-        vap_info->u.bss_info.am_config.npc.speed_tier = 0;
+        vap_info->u.bss_info.am_config.npc.speed_tier = 2;
         vap_info->u.bss_info.mdu_enabled = false;
         wifi_hal_get_default_ssid(ssid, vap_info->vap_index);
         wifi_hal_get_default_keypassphrase(password, vap_info->vap_index);
